src/components/ml/requirements.txt
# Vision-Y ML Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# ML Dependencies
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Utility Dependencies
python-dotenv==1.0.0
requests==2.31.0